split update logic appropriately

+++ app/controllers/examinations_controller.rb
@@ -13,22 +13,30 @@ class ExaminationsController < ApplicationController
 
 	#faster if batch load suppliers
 	def submit_examinations
-		datadump_id = Tag.find_by_name("datadump").id
-		params[:suppliers].each do |s_id,v|
-			supplier = Supplier.find(s_id)
-			if v == "duplicate"
-				supplier.destroy
-			elsif v == "not_duplicate"
-				supplier.update_attributes({name: params[:supplier_names][s_id]})
-				supplier.create_or_update_address({ country: params[:supplier_countries][s_id], 
-																						state: params[:supplier_states][s_id],
-																						zip: params[:supplier_zips][s_id]
-																					})
-				supplier.remove_tags(datadump_id)
+		note = "Didn't submit to a model name. Error in code."
+		if params[:model_examined] == "supplier"
+			datadump_id = Tag.find_by_name("datadump").id
+			if params[:suppliers]
+				params[:suppliers].each do |s_id,v|
+					supplier = Supplier.find(s_id)
+					if v == "duplicate"
+						supplier.destroy
+					elsif v == "not_duplicate"
+						supplier.update_attributes({name: params[:supplier_names][s_id]})
+						supplier.create_or_update_address({ country: params[:supplier_countries][s_id], 
+																								state: params[:supplier_states][s_id],
+																								zip: params[:supplier_zips][s_id]
+																							})
+						supplier.remove_tags(datadump_id)
+					end
+				end
 			end
-		end
+			note = "Suppliers submitted"
+		elsif params[:model_examined] == "review"
+			note = "Reviews submitted"
+		end 
 
-		redirect_to setup_examinations_path, notice: "Examinations submitted."
+		redirect_to setup_examinations_path(params[:model_examined]), notice: note
 	end
 
 end
